Consolidate event bus types
