--- src/@types/inputs-event.ts
@@ -1,6 +1,0 @@
-/**
- * Keys for input related event.
- *  - `keyboard/init`, for when a keyboard is first detected.
- *  - `gamepad/init`, for when a specific gamepad is first used in a session.
- */
-export type InputsEvent = "keyboard/init" | "gamepad/init";
