--- src/@types/touch-controls-event.ts
@@ -1,6 +1,0 @@
-export type TouchControlsEvent =
-  | "touchControls/move/start"
-  | "touchControls/move/end"
-  | "touchControls/move/save"
-  | "touchControls/move/cancel"
-  | "touchControls/move/reset";
