+++ src/event-bus.ts
@@ -1,7 +1,4 @@
-import type { InputsEvent } from "#types/inputs-event";
-import type { LanguageEvent } from "#types/language";
-import type { SettingsEvent } from "#types/settings";
-import type { TouchControlsEvent } from "#types/touch-controls-event";
+import type { InputsEvent, LanguageEvent, SettingsEvent, TouchControlsEvent } from "#types/event-bus-types";
 import Phaser from "phaser";
 
 type EventName = SettingsEvent | InputsEvent | TouchControlsEvent | LanguageEvent;
